Add files via upload
This is a student record management system project, which enables you to create, read, modify, and delete student Records. It utilizes file handling and shows the class and object of the programming language. It is a very effective and simple understanding of beginners. This is a simple mini project in C++, built as a console application without using any graphics features.